+++ Opencast/App/AppDelegate.swift
@@ -2,7 +2,6 @@ import AppKit
 
 final class AppDelegate: NSObject, NSApplicationDelegate {
     func applicationDidFinishLaunching(_ notification: Notification) {
-        if UpdateInstaller.applyIfRequested() { return }
         AppCore.shared.start()
     }